создание отдю шаблона
