+++ src/tools/destination-tools.js
@@ -1,0 +1,5 @@
+const generateDestList = (destination) => `${destination.map((elem) => `<option value="${elem.name}"></option>`).join('')}`;
+
+const getCurrentDestination = (destinationId,destination) => destination.find((element) => element.id === destinationId);
+
+export {generateDestList,getCurrentDestination};
